moved Actor/Schema models to shared/models
