+++ trustregistry/registry/registry_actors.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 from shared.log_config import get_logger
 from trustregistry import crud
 from trustregistry.db import get_db
-from trustregistry.schemas import Actor
+from shared.models.trustregistry import Actor
 
 logger = get_logger(__name__)
